+++ src/main/java/com/vh/repository/impl/OtherStaffDAO.java
@@ -79,7 +79,7 @@ public class OtherStaffDAO extends WorkerDAO{
                     resultSet.getString("name"),
                     resultSet.getDate("birthday"),
                     resultSet.getDate("start_date"),
-                    resultSet.getInt("salary"),
+                    resultSet.getLong("salary"),
                     RoleType.valueOf(resultSet.getString("role")),
                     DepartmentType.valueOf(resultSet.getString("department")),
                     resultSet.getString("description")));
